Attempt to fix tests on TravisCI and add Collision output
Our issue does not seem to be a Travis issue but more likely a Testbench one related to the fact that we copy files in the Laravel project it uses. It seems like added files are not being picked up on first run. To reproduce locally, rm -rf vendor/orchestra/testbench-core && composer install.
