+++ phpunit.xml
@@ -6,6 +6,7 @@
          convertErrorsToExceptions="true"
          convertNoticesToExceptions="true"
          convertWarningsToExceptions="true"
+         printerClass="NunoMaduro\Collision\Adapters\Phpunit\Printer"
          processIsolation="false"
          stopOnError="false"
          stopOnFailure="false"
