+++ tests/integration/ModulesAuthorsTest.php
@@ -72,6 +72,24 @@ class ModulesAuthorsTest extends ModulesTestBase
         $this->request('/twill/categories')->assertStatus(200);
     }
 
+    public function testCanSearchString()
+    {
+        $this->createAuthor(3);
+
+        $this->ajax("/twill/search?search={$this->name_en}")->assertStatus(200);
+
+        $this->assertJson($this->content());
+
+        $result = json_decode($this->content(), true);
+
+        $this->assertGreaterThan(0, count($result));
+
+        $this->assertEquals(
+            $this->now->format('Y-m-d\TH:i:s+00:00'),
+            $result[0]['date']
+        );
+    }
+
     public function testCanStartRestoringRevision()
     {
         $this->createAuthor();
