+++ tests/stubs/modules/authors/twill.php
@@ -38,7 +38,7 @@ return [
                             'vendor/area17/twill/src/Commands/stubs/blocks'
                         ),
                         'source' =>
-                            A17\Twill\Services\Blocks\Block::SOURCE_TWILL,
+                        A17\Twill\Services\Blocks\Block::SOURCE_TWILL,
                     ],
                     [
                         'path' => resource_path('views/admin/blocks'),
@@ -56,7 +56,7 @@ return [
                             'vendor/area17/twill/src/Commands/stubs/repeaters'
                         ),
                         'source' =>
-                            A17\Twill\Services\Blocks\Block::SOURCE_TWILL,
+                        A17\Twill\Services\Blocks\Block::SOURCE_TWILL,
                     ],
                 ],
 
@@ -74,5 +74,37 @@ return [
                 'repeaters' => resource_path('views/admin/repeaters'),
             ],
         ],
+    ],
+    'dashboard' => [
+        'modules' => [
+            App\Models\Author::class => [
+                'name' => 'authors', // module name
+                'label' => 'authors', // optional, if the name of your module above does not work as a label
+                'label_singular' => 'author', // optional, if the automated singular version of your name/label above does not work as a label
+                'routePrefix' => 'personnel', // optional, if the module is living under a specific routes group
+                'count' => true, // show total count with link to index of this module
+                'create' => true, // show link in create new dropdown
+                'activity' => true, // show activities on this module in actities list
+                'draft' => true, // show drafts of this module for current user
+                'search' => true, // show results for this module in global search
+                'search_fields' => ['name'],
+            ],
+
+            App\Models\Category::class => [
+                'name' => 'categories', // module name
+                'label' => 'categories', // optional, if the name of your module above does not work as a label
+                'label_singular' => 'category', // optional, if the automated singular version of your name/label above does not work as a label
+                'routePrefix' => '', // optional, if the module is living under a specific routes group
+                'count' => true, // show total count with link to index of this module
+                'create' => true, // show link in create new dropdown
+                'activity' => true, // show activities on this module in actities list
+                'draft' => true, // show drafts of this module for current user
+                'search' => true, // show results for this module in global search
+            ],
+        ],
+
+        'analytics' => ['enabled' => false],
+
+        'search_endpoint' => 'admin.search',
     ],
 ];
